Bunch of changes in preparation for dropping 32-bit limits and bytecode level 9.
- Most functions now take int64_t. There's still some use of int32_t,
  but ultimately it'll probably be removed over time.

- Added bytecode level 9 support, which currently just adds typed
  values. Bytecode 8 still technically supported, but will probably be
  dropped as bytecode 9 is fleshed out.

- Data values are now defined as 64-bit signed integers. Relocations
  handle this accordingly.

- The unit loader probably needs to be rewritten again, because it's
  kind of a mess.

+++ _types.h
@@ -28,7 +28,7 @@ struct vm_find_result {
   T value;
 };
 
-using vm_found_fn_t = vm_find_result<int32_t>;
+using vm_found_fn_t = vm_find_result<int64_t>;
 using vm_bound_fn_t = vm_found_fn_t;
