+++ vm_state.h
@@ -40,15 +40,15 @@ enum vm_memblock_flags : uint32_t
   VM_MEM_SOURCE_DATA = VM_MEM_STATIC | VM_MEM_READABLE,
 };
 
-constexpr int32_t VM_NULL_BLOCK = 0;
+constexpr int64_t VM_NULL_BLOCK = 0;
 
 
 class vm_state
 {
   struct memblock
   {
-    int32_t size;
-    uint32_t flags;
+    int64_t size;
+    uint32_t flags; // todo: reorder to remove padding bytes
     void *block;
   };
 
@@ -74,13 +74,13 @@ class vm_state
   thread_stores_t _threads {};
   callbacks_t _callbacks {};
   memblock_map_t _blocks {};
-  int32_t _block_counter = 1;
+  int64_t _block_counter = 1;
 
-  int32_t unused_block_id();
+  int64_t unused_block_id();
   void release_all_memblocks() noexcept;
 
   vm_unit _unit;
-  int32_t _source_size;
+  int64_t _source_size;
 
   void prepare_unit();
 
@@ -99,22 +99,22 @@ public:
   void set_unit(vm_unit &&unit);
 
 private:
-  bool check_block_bounds(int32_t block_id, int32_t offset, int32_t size) const;
+  bool check_block_bounds(int64_t block_id, int64_t offset, int64_t size) const;
   void load_thread(thread_pointer_t &&thread);
-  void destroy_thread(int32_t thread_index);
+  void destroy_thread(int64_t thread_index);
 
-  int32_t realloc_block_with_flags(int32_t block_id, int32_t size, uint32_t flags);
+  int64_t realloc_block_with_flags(int64_t block_id, int64_t size, uint32_t flags);
   // Returns the block for the given ID -- does not do flag checking of any kind.
-  found_memblock_t get_block_info(int32_t block_id) const;
+  found_memblock_t get_block_info(int64_t block_id) const;
 
 public:
-  int32_t realloc_block(int32_t block, int32_t size);
-  int32_t alloc_block(int32_t size) { return realloc_block(0, size); }
-  void free_block(int32_t block_id);
-  int32_t block_size(int32_t block_id) const;
-  int32_t duplicate_block(int32_t block_id);
-  void *get_block(int32_t block_id, uint32_t permissions);
-  const void *get_block(int32_t block_id, uint32_t permissions) const;
+  int64_t realloc_block(int64_t block, int64_t size);
+  int64_t alloc_block(int64_t size) { return realloc_block(0, size); }
+  void free_block(int64_t block_id);
+  int64_t block_size(int64_t block_id) const;
+  int64_t duplicate_block(int64_t block_id);
+  void *get_block(int64_t block_id, uint32_t permissions);
+  const void *get_block(int64_t block_id, uint32_t permissions) const;
 
   vm_found_fn_t find_function_pointer(const char *name) const;
 
@@ -124,8 +124,8 @@ public:
     return bind_callback(name, std::strlen(name), function, context);
   }
 
-  vm_thread &thread_by_index(int32_t thread_index);
-  vm_thread const &thread_by_index(int32_t thread_index) const;
+  vm_thread &thread_by_index(int64_t thread_index);
+  vm_thread const &thread_by_index(int64_t thread_index) const;
 
   vm_thread &make_thread(size_t stack_size = 8192);
   vm_thread &fork_thread(vm_thread const &thread);
